+++ packages/forms/src/index.ts
@@ -4,8 +4,8 @@ export * from './enhanced/FormikEnhanced';
 export * from './enhanced/FormsProvider';
 export * from './enhanced/useFormikEnhanced';
 export * from './phone/FormikPhoneField';
+export * from './radio/FormikRadioFieldCard';
 export * from './radio/FormikRadioGroupField';
-export * from './radioCard/FormikRadioCardField';
 export * from './text/constants';
 export * from './text/FormikCurrencyField';
 export * from './text/FormikMaxLengthTextField';
